* commit 'f069774faa33f03bc55eaf031e97772a342488d4':
  adding check for function exists
  php7.4 changes for release/15.0

+++ amp_conf/htdocs/admin/libraries/view.functions.php
@@ -11,11 +11,12 @@ function frameworkPasswordCheck() {
   //
 
 	// send error if magic_quotes_gpc is enabled on this system as much of the code base assumes not
-	//
-	if(get_magic_quotes_gpc()) {
-		$nt->add_error('core', 'MQGPC', _("Magic Quotes GPC"), _("You have magic_quotes_gpc enabled in your php.ini, http or .htaccess file which will cause errors in some modules. FreePBX expects this to be off and runs under that assumption"));
-	} else {
-		$nt->delete('core', 'MQGPC');
+	if(function_exists('get_magic_quotes_gpc')){
+		if(get_magic_quotes_gpc()) {
+			$nt->add_error('core', 'MQGPC', _("Magic Quotes GPC"), _("You have magic_quotes_gpc enabled in your php.ini, http or .htaccess file which will cause errors in some modules. FreePBX expects this to be off and runs under that assumption"));
+		} else {
+			$nt->delete('core', 'MQGPC');
+		}
 	}
 }
